requirements.txt
Flask==2.3.2
torch==2.1.0
torchvision==0.15.0
PILLOW==9.6.0
numpy==1.23.5
